create main py
It aint an library rn. its just a window with a rectangle falling and bouncing off. Also, when you shake the window, the rectangle moves too. 

Also for those who are gonna say its a cube not a rectangle, shut the f**k up a cube is also an rectangle so yea i won muhaha